Fix qt5 event loop
