+++ news/2 Fixes/8322.md
@@ -1,0 +1,3 @@
+Get ```%matplotlib qt5``` working again. 
+
+QT5's event loop is slightly different and can't handle concurrent requests so make sure to wait for a request before issuing another.
