+++ src/client/datascience/jupyter/kernels/kernel.ts
@@ -44,6 +44,7 @@ import {
     KernelSocketInformation
 } from '../../types';
 import {
+    executeSilently,
     getDisplayNameOrNameOfKernelConnection,
     getSysInfoReasonHeader,
     isPythonKernelConnection,
@@ -899,101 +900,6 @@ export class Kernel implements IKernel {
         }
         await executeSilently(this.notebook.session, code);
     }
-}
-
-export function executeSilentlySync(session: IJupyterSession, code: string) {
-    traceInfo(
-        `Executing (and forget) (status ${session.status}) silently Code = ${code
-            .substring(0, 100)
-            .splitLines()
-            .join('\\n')}`
-    );
-    session.requestExecute(
-        {
-            code: code.replace(/\r\n/g, '\n'),
-            silent: false,
-            stop_on_error: false,
-            allow_stdin: true,
-            store_history: false
-        },
-        true
-    );
-}
-
-export async function executeSilently(session: IJupyterSession, code: string): Promise<nbformat.IOutput[]> {
-    traceInfo(
-        `Executing (status ${session.status}) silently Code = ${code.substring(0, 100).splitLines().join('\\n')}`
-    );
-    // eslint-disable-next-line @typescript-eslint/no-require-imports
-    const jupyterLab = require('@jupyterlab/services') as typeof import('@jupyterlab/services');
-
-    const request = session.requestExecute(
-        {
-            code: code.replace(/\r\n/g, '\n'),
-            silent: false,
-            stop_on_error: false,
-            allow_stdin: true,
-            store_history: false
-        },
-        true
-    );
-    const outputs: nbformat.IOutput[] = [];
-    request.onIOPub = (msg) => {
-        if (jupyterLab.KernelMessage.isStreamMsg(msg)) {
-            traceInfoIfCI(`Got io pub message (stream), ${msg.content.text.substr(0, 100).splitLines().join('\\n')}`);
-            if (
-                outputs.length > 0 &&
-                outputs[outputs.length - 1].output_type === 'stream' &&
-                outputs[outputs.length - 1].name === msg.content.name
-            ) {
-                const streamOutput = outputs[outputs.length - 1] as nbformat.IStream;
-                streamOutput.text += msg.content.text;
-            } else {
-                const streamOutput: nbformat.IStream = {
-                    name: msg.content.name,
-                    text: msg.content.text,
-                    output_type: 'stream'
-                };
-                outputs.push(streamOutput);
-            }
-        } else if (jupyterLab.KernelMessage.isExecuteResultMsg(msg)) {
-            traceInfoIfCI(`Got io pub message (execresult)}`);
-            const output: nbformat.IExecuteResult = {
-                data: msg.content.data,
-                execution_count: msg.content.execution_count,
-                metadata: msg.content.metadata,
-                output_type: 'execute_result'
-            };
-            outputs.push(output);
-        } else if (jupyterLab.KernelMessage.isDisplayDataMsg(msg)) {
-            traceInfoIfCI(`Got io pub message (displaydata)}`);
-            const output: nbformat.IDisplayData = {
-                data: msg.content.data,
-                metadata: msg.content.metadata,
-                output_type: 'display_data'
-            };
-            outputs.push(output);
-        } else if (jupyterLab.KernelMessage.isErrorMsg(msg)) {
-            traceInfoIfCI(
-                `Got io pub message (error), ${msg.content.ename},${
-                    msg.content.evalue
-                }, ${msg.content.traceback.join().substring(0, 100)}}`
-            );
-            const output: nbformat.IError = {
-                ename: msg.content.ename,
-                evalue: msg.content.evalue,
-                traceback: msg.content.traceback,
-                output_type: 'error'
-            };
-            outputs.push(output);
-        } else {
-            traceInfoIfCI(`Got io pub message (${msg.header.msg_type})`);
-        }
-    };
-    await request.done;
-    traceInfo(`Executing silently Code (completed) = ${code.substring(0, 100).splitLines().join('\\n')}`);
-
-    return outputs;
 }
 
 /**
